+++ tests/integration/test_variables.py
@@ -31,29 +31,24 @@ def test_agent_pass_batch_payload_size_variable(mock_client_init):
     result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
     assert result == 0  # the test successfully passed
 
-    payload_variable = "log_batch_payload_size"
+    payload_variable = "log_batch_payload_limit"
     assert payload_variable in mock_client_init.call_args_list[0][1]
     assert mock_client_init.call_args_list[0][1][payload_variable] == payload_size
 
 
-@mock.patch(REQUESTS_SERVICE)
-def test_agent_pass_launch_uuid_variable(mock_requests_init):
+@mock.patch(REPORT_PORTAL_SERVICE)
+def test_agent_pass_launch_uuid_variable(mock_client_init):
     variables = utils.DEFAULT_VARIABLES.copy()
     test_launch_id = "my_test_launch"
     variables["RP_LAUNCH_UUID"] = test_launch_id
     result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
     assert result == 0  # the test successfully passed
 
-    mock_requests = mock_requests_init.return_value
-    assert mock_requests.post.call_count == 3
-    suite_start = mock_requests.post.call_args_list[0]
-    assert suite_start[0][0].endswith("/item")
-    assert suite_start[1]["json"]["launchUuid"] == test_launch_id
+    assert "launch_uuid" in mock_client_init.call_args_list[0][1]
+    assert mock_client_init.call_args_list[0][1]["launch_uuid"] == test_launch_id
 
 
-@pytest.mark.parametrize(
-    "variable, warn_num", [("RP_PROJECT", 1), ("RP_API_KEY", 2), ("RP_ENDPOINT", 1), ("RP_LAUNCH", 1)]
-)
+@pytest.mark.parametrize("variable, warn_num", [("RP_PROJECT", 1), ("RP_ENDPOINT", 1), ("RP_LAUNCH", 1)])
 @mock.patch(REPORT_PORTAL_SERVICE)
 def test_no_required_variable_warning(mock_client_init, variable, warn_num):
     variables = utils.DEFAULT_VARIABLES.copy()
@@ -101,53 +96,15 @@ def test_rp_api_key(mock_client_init):
         assert len(filter_agent_calls(w)) == 0
 
 
-@mock.patch(REPORT_PORTAL_SERVICE)
-def test_rp_uuid(mock_client_init):
-    api_key = "rp_api_key"
-    variables = dict(utils.DEFAULT_VARIABLES)
-    del variables["RP_API_KEY"]
-    variables.update({"RP_UUID": api_key}.items())
-
-    with warnings.catch_warnings(record=True) as w:
-        result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-        assert int(result) == 0, "Exit code should be 0 (no errors)"
-
-        assert mock_client_init.call_count == 1
-
-        constructor_args = mock_client_init.call_args_list[0][1]
-        assert constructor_args["api_key"] == api_key
-        assert len(filter_agent_calls(w)) == 1
-
-
-@mock.patch(REPORT_PORTAL_SERVICE)
-def test_rp_api_key_priority(mock_client_init):
-    api_key = "rp_api_key"
-    variables = dict(utils.DEFAULT_VARIABLES)
-    variables.update({"RP_API_KEY": api_key, "RP_UUID": "rp_uuid"}.items())
-
-    with warnings.catch_warnings(record=True) as w:
-        result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-        assert int(result) == 0, "Exit code should be 0 (no errors)"
-
-        assert mock_client_init.call_count == 1
-
-        constructor_args = mock_client_init.call_args_list[0][1]
-        assert constructor_args["api_key"] == api_key
-        assert len(filter_agent_calls(w)) == 0
-
-
-@mock.patch(REPORT_PORTAL_SERVICE)
-def test_rp_api_key_empty(mock_client_init):
+def test_rp_api_key_empty():
     api_key = ""
     variables = dict(utils.DEFAULT_VARIABLES)
     variables.update({"RP_API_KEY": api_key}.items())
 
-    with warnings.catch_warnings(record=True) as w:
-        result = utils.run_robot_tests(["examples/simple.robot"], variables=variables)
-        assert int(result) == 0, "Exit code should be 0 (no errors)"
-
-        assert mock_client_init.call_count == 0
-        assert len(filter_agent_calls(w)) == 2
+    with pytest.warns(Warning) as w:
+        utils.run_robot_tests(["examples/simple.robot"], variables=variables)
+        assert len(w) == 1
+        assert "Authentication credentials are required." in w[0].message.args[0]
 
 
 @mock.patch(REPORT_PORTAL_SERVICE)